requirements.txt.txt
django == 3.2.16
pandas == 1.3.5
django-import-export == 3.2.0